+++ python/GafferImageUITest/ImageGadgetTest.py
@@ -37,6 +37,8 @@
 import unittest
 import imath
 
+import IECore
+
 import Gaffer
 import GafferUI
 import GafferUITest
@@ -95,7 +97,15 @@ class ImageGadgetTest( GafferUITest.TestCase ) :
 
 		w.setVisible( True )
 
-		self.waitForIdle( 1000 )
+		# If this computer doesn't support floating point textures, the ImageGadget will warn about
+		# this the first time it tries to render.  Don't fail because of this
+		with IECore.CapturingMessageHandler() as mh :
+			self.waitForIdle( 1000 )
+
+		if len( mh.messages ):
+			self.assertEqual(  len( mh.messages ), 1 )
+			self.assertEqual( mh.messages[0].context, "ImageGadget" )
+			self.assertEqual( mh.messages[0].message, "Could not find supported floating point texture format in OpenGL.  GPU image viewer path will be low quality, recommend switching to CPU display transform, or resolving graphics driver issue." )
 
 		del g, w
 		del s